feat: Implement HLS video processing module with Express
- Added main application logic in `src/app.js` for handling HLS video processing.
- Integrated ffmpeg for video transcoding and segmenting.
- Created database connection pool in `src/db.js` for MySQL.
- Added Jest testing framework with initial test cases for bitrate calculation, language info retrieval, and hash service calls.
- Implemented mock files for testing purposes.
- Added endpoints for HLS processing and cancellation.
- Updated package.json with necessary dependencies and scripts for testing and coverage.
- Created necessary mock files for database and ffmpeg in `__mocks__` directory.
- Added .gitignore to exclude node_modules.
