+++ modules/generator-hls/__mocks__/db.js
@@ -1,0 +1,1 @@
+module.exports = { query: jest.fn().mockResolvedValue([]) };
