+++ modules/generator-hls/src/db.js
