+++ modules/generator-hls/tests/jest.setup.js
@@ -1,0 +1,2 @@
+const mock = require('mock-fs');
+afterEach(() => mock.restore());
